LUCENE-5477: add near-real-time add/update/refresh to AnalyzingInfixSuggester
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/branch_4x@1572997 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/codecs/src/java/org/apache/lucene/codecs/simpletext/SimpleTextFieldsReader.java
@@ -568,7 +568,7 @@ class SimpleTextFieldsReader extends FieldsProducer {
           termCount++;
         }
       }
-      docCount = (int) visitedDocs.cardinality();
+      docCount = visitedDocs.cardinality();
       fst = b.finish();
       /*
       PrintStream ps = new PrintStream("out.dot");
